refactor and fix 'get retrieval tool for docset'
